+++ src/pages/benefit/components/PaginationButton.types.ts
@@ -1,0 +1,5 @@
+export interface PaginationButtonProps {
+  onClick: () => void;
+  disabled: boolean;
+  direction: string;
+}
