+++ src/pages/benefit/types/Pagination.types.ts
@@ -1,0 +1,5 @@
+export interface PaginationProps {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (page: number) => void;
+}
